+++ R/sentomodel.R
@@ -227,7 +227,7 @@ ctr_model <- function(model = c("gaussian", "binomial", "multinomial"), type = c
 #'
 #' @return If \code{ctr$do.iter = TRUE}, a \code{sentomodeliter} object which is a \code{list} containing:
 #' \item{models}{all sparse regressions, i.e., separate \code{sentomodel} objects as above, as a \code{list} with as names the
-#' dates from the perspective of the sentiment measures at which predictions for performance measurement are carried out.}
+#' dates from the perspective of the sentiment measures at which the out-of-sample predictions are carried out.}
 #' \item{alphas}{calibrated alphas.}
 #' \item{lambdas}{calibrated lambdas.}
 #' \item{performance}{a \code{data.frame} with performance-related measures, being "\code{RMSFE}" (root mean squared